portfolio_management/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via -r requirements.in
aiosignal==1.4.0
    # via aiohttp
asgiref==3.9.1
    # via
    #   -r requirements.in
    #   channels
    #   django
    #   django-cors-headers
attrs==25.3.0
    # via aiohttp
babel==2.17.0
    # via -r requirements.in
beautifulsoup4==4.13.5
    # via
    #   -r requirements.in
    #   yfinance
cachetools==5.5.2
    # via tinkoff-investments
certifi==2025.8.3
    # via
    #   curl-cffi
    #   requests
cffi==1.17.1
    # via
    #   cryptography
    #   curl-cffi
channels==4.3.1
    # via -r requirements.in
chardet==5.2.0
    # via -r requirements.in
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via uvicorn
colorama==0.4.6
    # via click
cryptography==45.0.7
    # via
    #   -r requirements.in
    #   pyopenssl
curl-cffi==0.13.0
    # via yfinance
deprecation==2.1.0
    # via tinkoff-investments
django==5.2.6
    # via
    #   -r requirements.in
    #   channels
    #   django-cors-headers
    #   django-debug-toolbar
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==4.7.0
    # via -r requirements.in
django-debug-toolbar==6.0.0
    # via -r requirements.in
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
fake-useragent==2.2.0
    # via -r requirements.in
filelock==3.19.1
    # via -r requirements.in
frozendict==2.4.6
    # via yfinance
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fuzzywuzzy==0.18.0
    # via -r requirements.in
grpcio==1.74.0
    # via tinkoff-investments
h11==0.16.0
    # via uvicorn
idna==3.10
    # via
    #   requests
    #   yarl
levenshtein==0.27.1
    # via python-levenshtein
multidict==6.6.4
    # via
    #   aiohttp
    #   yarl
multitasking==0.0.12
    # via yfinance
networkx==3.5
    # via -r requirements.in
numpy==2.3.2
    # via
    #   -r requirements.in
    #   pandas
    #   yfinance
packaging==25.0
    # via deprecation
pandas==2.3.2
    # via
    #   -r requirements.in
    #   yfinance
peewee==3.18.2
    # via yfinance
platformdirs==4.4.0
    # via yfinance
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
protobuf==4.25.8
    # via
    #   tinkoff-investments
    #   yfinance
pycparser==2.22
    # via cffi
pyjwt==2.10.1
    # via djangorestframework-simplejwt
pyopenssl==25.1.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   pandas
    #   tinkoff-investments
python-levenshtein==0.27.1
    # via -r requirements.in
pytz==2025.2
    # via
    #   pandas
    #   yfinance
pyxirr==0.10.7
    # via -r requirements.in
rapidfuzz==3.14.0
    # via levenshtein
requests==2.32.5
    # via yfinance
simplejson==3.20.1
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
soupsieve==2.8
    # via beautifulsoup4
sqlparse==0.5.3
    # via
    #   django
    #   django-debug-toolbar
structlog==25.4.0
    # via -r requirements.in
tinkoff==0.1.1
    # via
    #   -r requirements.in
    #   tinkoff-investments
tinkoff-investments==0.2.0b116
    # via -r requirements.in
typing-extensions==4.15.0
    # via beautifulsoup4
tzdata==2025.2
    # via
    #   django
    #   pandas
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via -r requirements.in
websockets==15.0.1
    # via yfinance
yarl==1.20.1
    # via aiohttp
yfinance==0.2.66
    # via -r requirements.in
zstandard==0.24.0
    # via -r requirements.in
